# chore: Optimistic E3 Execution for Poseidon2 Chip