… extensions/native/circuit/src/poseidon2/chip.rs