@jonathanpwang 2025-07-25
Please provide an update on the measured time in insn/s of just verifybatch before and after this PR